:green_heart: Add test
